+++ src/main/java/gtclassic/common/GTWorldGen.java
@@ -23,23 +23,34 @@ public class GTWorldGen implements IWorldGenerator {
 			IChunkProvider chunkProvider) {
 		Biome biomegenbase = world.getBiome(new BlockPos(chunkX * 16 + 16, 128, chunkZ * 16 + 16));
 		// Any Biome
-		GTOreGenerator.generateBasicVein(GTBlocks.oreIridium, GTConfig.generation.iridiumGenerate, GTConfig.generation.iridiumSize, GTConfig.generation.iridiumWeight, 0, 128, Blocks.STONE, world, random, chunkX, chunkZ);
+		GTOreGenerator.generateBasicVein(GTBlocks.oreIridium, GTConfig.generation.iridiumGenerate,
+				GTConfig.generation.iridiumSize, GTConfig.generation.iridiumWeight, 0, 128, Blocks.STONE, world, random,
+				chunkX, chunkZ);
 		// Jungle Biomes
 		if (BiomeDictionary.hasType(biomegenbase, Type.JUNGLE)) {
-			GTOreGenerator.generateBasicVein(GTBlocks.oreSheldonite, GTConfig.generation.sheldoniteGenerate, GTConfig.generation.sheldoniteSize, GTConfig.generation.sheldoniteWeight, 10, 30, Blocks.STONE, world, random, chunkX, chunkZ);
+		GTOreGenerator.generateBasicVein(GTBlocks.oreSheldonite, GTConfig.generation.sheldoniteGenerate,
+				GTConfig.generation.sheldoniteSize, GTConfig.generation.sheldoniteWeight, 10, 30, Blocks.STONE, world,
+				random, chunkX, chunkZ);
 		}
 		// Hot Biomes
-		if (BiomeDictionary.hasType(biomegenbase, Type.HOT)) {
-			GTOreGenerator.generateBasicVein(GTBlocks.oreRuby, GTConfig.generation.rubyGenerate, GTConfig.generation.rubySize, GTConfig.generation.rubyWeight, 0, 48, Blocks.STONE, world, random, chunkX, chunkZ);
+		if (BiomeDictionary.hasType(biomegenbase, Type.HOT) || BiomeDictionary.hasType(biomegenbase, Type.DRY) || BiomeDictionary.hasType(biomegenbase, Type.SAVANNA) || BiomeDictionary.hasType(biomegenbase, Type.WASTELAND) ) {
+			GTOreGenerator.generateBasicVein(GTBlocks.oreRuby, GTConfig.generation.rubyGenerate,
+					GTConfig.generation.rubySize, GTConfig.generation.rubyWeight, 0, 48, Blocks.STONE, world, random,
+					chunkX, chunkZ);
 		}
-		// Ocean Biomes
-		if (BiomeDictionary.hasType(biomegenbase, Type.OCEAN) || BiomeDictionary.hasType(biomegenbase, Type.BEACH)) {
-			GTOreGenerator.generateBasicVein(GTBlocks.oreSapphire, GTConfig.generation.sapphireGenerate, GTConfig.generation.sapphireSize, GTConfig.generation.sapphireWeight, 0, 48, Blocks.STONE, world, random, chunkX, chunkZ);
+		// Wet Biomes
+		if (BiomeDictionary.hasType(biomegenbase, Type.OCEAN) || BiomeDictionary.hasType(biomegenbase, Type.BEACH)
+				|| BiomeDictionary.hasType(biomegenbase, Type.RIVER)) {
+			GTOreGenerator.generateBasicVein(GTBlocks.oreSapphire, GTConfig.generation.sapphireGenerate,
+					GTConfig.generation.sapphireSize, GTConfig.generation.sapphireWeight, 0, 48, Blocks.STONE, world,
+					random, chunkX, chunkZ);
 		}
 		// Forest or Plains Biomes
 		if (BiomeDictionary.hasType(biomegenbase, Type.FOREST)
 				|| (BiomeDictionary.hasType(biomegenbase, Type.PLAINS))) {
-			GTOreGenerator.generateBasicVein(GTBlocks.oreBauxite, GTConfig.generation.bauxiteGenerate, GTConfig.generation.bauxiteSize, GTConfig.generation.bauxiteWeight, 50, 120, Blocks.STONE, world, random, chunkX, chunkZ);
+			GTOreGenerator.generateBasicVein(GTBlocks.oreBauxite, GTConfig.generation.bauxiteGenerate,
+					GTConfig.generation.bauxiteSize, GTConfig.generation.bauxiteWeight, 50, 120, Blocks.STONE, world,
+					random, chunkX, chunkZ);
 		}
 		if (world.provider.getDimensionType().equals(DimensionType.OVERWORLD)) {
 			for (Block block : GTBedrockOreHandler.getBedrockOreMap().keySet()) {